+++ components/cli/src/domain/use-cases/get-static-token.ts
@@ -1,0 +1,50 @@
+import type { Envelope, QueryHandlerDefinition } from 'missive.js';
+import type { PimCredentials } from '../contracts/models/credentials';
+import type { createClient } from '@crystallize/js-api-client';
+
+type Deps = {
+    createCrystallizeClient: typeof createClient;
+};
+
+type Query = {
+    tenantIdentifier: string;
+    credentials: PimCredentials;
+};
+
+export type GetStaticAuthTokenHandlerDefinition = QueryHandlerDefinition<
+    'GetStaticAuthToken',
+    Query,
+    Awaited<ReturnType<typeof handler>>
+>;
+
+const handler = async (envelope: Envelope<Query>, deps: Deps) => {
+    const { createCrystallizeClient } = deps;
+    const { tenantIdentifier, credentials } = envelope.message;
+    const client = createCrystallizeClient({
+        tenantIdentifier: tenantIdentifier,
+        accessTokenId: credentials.ACCESS_TOKEN_ID,
+        accessTokenSecret: credentials.ACCESS_TOKEN_SECRET,
+    });
+
+    const results = await client.nextPimApi(`#graphql
+        query {
+            tenant {
+                ... on Tenant {
+                    staticAuthToken
+                }
+                ... on BasicError {
+                    error: message
+                }
+            }
+        }
+    `);
+
+    if (results.tenant.error) {
+        throw new Error(results.tenant.error);
+    }
+    return {
+        token: results.tenant.staticAuthToken,
+    };
+};
+
+export const createGetStaticAuthTokenHandler = (deps: Deps) => (query: Envelope<Query>) => handler(query, deps);